talk-to-data/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
groq
pandas
python-dotenv
numpy
reportlab
